Fix command types.
Currently, the redis-benchmark is too slow because we generate the
random data directly into the writehandler. This adds more work for the
CPU with directly proportional with the number of requests. If we have
more requests with smaller data this adds more overhead than having
bigger data with less requests. However, putting bigger data also adds a
it of overhead because the radom generation will have to generate bigger
data.

In the near future I need to fix this and maybe create the data offline
and use a signal handler to start the test

Change-Id: Ieef0c8f54a9dbf2ebb0139adb6bc247906d7e77f