Start stop containers instead of redeploy.
Start stop allows us to monitor CPU and Memory usage in Grafana. If we
redeploy containers, then Grafana would just see a new container and we
end up with hundreds of containers instead of monitoring just one.

Also fix a bug where traffic control was set on fixed device. Now we
take the grep the device based on IP.

Change-Id: I6413b9cdb1ce7a1362969f491dd120388fe9e896